enhance performance speed for new version

+++ app/src/main/java/com/bt_121shoppe/motorbike/Api/api/model/Item.java
@@ -69,7 +69,7 @@ public class Item {
         return id;
     }
 
-    public void setId(float id) {
+    public void setId(int id) {
         this.id = id;
     }
